import data transaksi

+++ data/src/main/java/me/skripsi/data/models/MappingExtension.kt
@@ -17,7 +17,7 @@ fun DataTransaksiEntity.toDataTransaksi() : DataTransaksi{
             namaBarang = namaBarang,
             golongan = golongan,
             stok = stok,
-            isDiskon = isDiskon,
+            isDiskon = isDiskon.toInt(),
             penjualan = penjualan,
             pembelian = pembelian
         )
